# backend/api/controller/warehouseTools.go
package controller

import (
	"net/http"
	"net/url"
	"strconv"
	"zartool/models"
	"zartool/repositories"

	"github.com/labstack/echo/v4"
)

// GetWareHouseTools godoc
//
//	@Summary        GetWareHouseTools
//	@Description    GetWareHouseTools
//	@Tags           zartool
//	@Accept         json
//	@Produce        json
//	@Security       JWT
//	@Param          page query int false "page"
//	@Param          page_size query int false "page_size"
//  @Success        200 {object} models.WarehouseToolsResponse
//	@Router         /warehouse-tools [get]
func (s *Controller) GetWareHouseTools(e echo.Context) error {
	var queries url.Values = e.QueryParams()

	page, _ := strconv.Atoi(queries.Get("page"))
	pageSize, _ := strconv.Atoi(queries.Get("page_size"))

	tools, meta, err := repositories.GetWareHouseTools(s.DB, page, pageSize)

	if err != nil {
		return e.JSON(http.StatusInternalServerError, models.ErrorResponse{Status: http.StatusInternalServerError, Message: "Internal server error"})
	}

	resp := models.WarehouseToolsResponse{
		Status:  http.StatusOK,
		Message: "Success",
		Data:    tools,
		Meta:    meta,
	}

	return e.JSON(http.StatusOK, resp)
}

// AddNewTools godoc
//	@Summary        AddNewTools
//	@Description    AddNewTools
//	@Tags           zartool
//	@Accept         json
//	@Produce        json
//	@Security       JWT
//	@Param          payload body models.WarehouseTools false "body"
//  @Success        200 {object} models.WarehouseToolsCreateResponse
//	@Router         /warehouse-tool/create [post]
func (s Controller) AddNewTools(e echo.Context) error {
	var newTool []models.WarehouseTools

	if err := e.Bind(&newTool); err != nil {
		return e.JSON(http.StatusInternalServerError, models.ErrorResponse{Status: http.StatusInternalServerError, Message: "Internal server error"})
	}

	err := repositories.AddNewTool(s.DB, newTool)

	if err != nil {
		return e.JSON(http.StatusBadRequest, models.ErrorResponse{Status: http.StatusBadRequest, Message: err.Error()})
	}

	resp := models.WarehouseToolsCreateResponse{
		Status:  http.StatusCreated,
		Message: "Succes",
		Data:    newTool,
	}

	return e.JSON(http.StatusCreated, resp)
}

// DeleteWarehouseTool godoc
//	@Summary        DeleteWarehouseTool
//	@Description    DeleteWarehouseTool
//	@Tags           zartool
//	@Accept         json
//	@Produce        json
//	@Security       JWT
//	@Param          id path int true "id"
//  @Success        200 {object} models.SuccessResponse
//	@Router         /warehouse-tool/delete/{id} [delete]
func (c Controller) DeleteWarehouseTool(e echo.Context) error {
	id, err := strconv.Atoi(e.Param("id"))

	if err != nil {
		return e.JSON(http.StatusInternalServerError, models.ErrorResponse{Status: http.StatusInternalServerError, Message: "Internal server error"})
	}

	err = repositories.DeleteWarehouseTool(c.DB, id)

	if err != nil {
		return e.JSON(http.StatusBadRequest, models.ErrorResponse{Status: http.StatusBadRequest, Message: err.Error()})
	}

	resp := models.SuccessResponse{
		Status:  http.StatusOK,
		Message: "Success",
	}

	return e.JSON(http.StatusOK, resp)
}

// UpdateWareHouseTool godoc
//	@Summary        UpdateWareHouseTool
//	@Description    UpdateWareHouseTool
//	@Tags           zartool
//	@Accept         json
//	@Produce        json
//	@Security       JWT
//	@Param          payload body models.WarehouseToolsUpdateResponse true "payload"
//  @Success        200 {object} models.SuccessResponse
//	@Router         /warehouse-tool/update/{id} [put]
func (c Controller) UpdateWareHouseTool(e echo.Context) error {
	var tool models.WarehouseTools

	if err := e.Bind(&tool); err != nil {
		return e.JSON(http.StatusInternalServerError, models.ErrorResponse{Status: http.StatusInternalServerError, Message: "Internal server error"})
	}

	err := repositories.UpdateWareHouseTool(c.DB, tool)

	if err != nil {
		return e.JSON(http.StatusBadRequest, models.ErrorResponse{Status: http.StatusBadRequest, Message: err.Error()})
	}

	resp := models.WarehouseToolsUpdateResponse{
		Status:  http.StatusOK,
		Message: "Success",
		Data:    tool,
	}

	return e.JSON(http.StatusOK, resp)
}